Refactor out abstract Mixer base class and move Brain- and Camera-related functionality into their own base classes
