+++ Runtime/Timeline/CinemachineBrainMixer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4da6402d2b53dc14bbfb942cb81c37e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
