+++ Runtime/Timeline/CinemachineCameraMixer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 472ccb5ccde1775418bca0452d2a0b5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
