+++ Runtime/Timeline/CinemachineMixerBase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 961772e8b1d75dd4a915fb8a236d5a8a
+guid: 189202d05a1e775488c8c02806dce93e
 timeCreated: 1489083733
 licenseType: Pro
 MonoImporter:
